pyproject.toml
[project]
name = "ssl-streaming-system"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyzmq (>=26.4.0,<27.0.0)",
    "protobuf (>=6.30.2,<7.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "playsound3 (>=3.2.3,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"